Introduce a criticalLoop() method
If a timer0 interrupt handler is called when a method of WiFi, EEPROM, ESP etc. method is being called, there will be a reset/stack dump because of an illegal instruction call. Introducing a criticalLoop method allows guards to be placed around that call, while leaving more innocuous code to be called at will.